Extract CountDownTimer as a separate component
